requirements.txt
django==2.1
django-ckeditor==5.6.1
pillow==5.3.0
gunicorn==19.7.1
djangorestframework==3.9.0
markdown==3.0.1
django-filter==2.0.0
django-cors-headers==2.4.0